% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pantompkins.R
\name{rpeaks_pan_tompkins}
\alias{rpeaks_pan_tompkins}
\title{Detect rpeaks using the pan-tompkins algorithm}
\usage{
rpeaks_pan_tompkins(
  ecg,
  sample_rate,
  integration_window = 0.15,
  refractory = 0.2
)
}
\arguments{
\item{ecg}{raw ecg data vector}

\item{sample_rate}{sampling rate in Hz of the ecg}

\item{integration_window}{size of the integration window in seconds}

\item{refractory}{refractory period in seconds (minimum time between peaks)}
}
\description{
Fast implementation of the pan-tompkins algorithm. The default parameters are
taken from the original pan and tompkins paper
}
\references{
Pan, J., & Tompkins, W. J. (1985). A real-time QRS detection algorithm. IEEE transactions on biomedical engineering, (3), 230-236.
}
